requirements.txt
streamlit>=1.30
pandas>=2.0
openpyxl>=3.1
folium>=0.15
streamlit-folium>=0.18
shapely>=2.0
psycopg2-binary>=2.9
SQLAlchemy>=2.0


